增加所有权重矩阵数据

+++ debugcode/src/types.h
@@ -1,7 +1,7 @@
 #ifndef TYPES
 #define TYPES
 
-typedef int FDATA_T;
+typedef double FDATA_T;
 typedef int LDATA_T;
 
 #endif
